% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSlidingWindow.R
\name{RSlidingWindow}
\alias{RSlidingWindow}
\title{RSlidingWindow}
\usage{
RSlidingWindow(x, length.sw, by.sw = 1, method.sw = "mean",
  keep.cut = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{length.sw}{An integer, length of the windows.}

\item{by.sw}{An integer, steps between each window [default: 1].}

\item{method.sw}{A string, name of the method to use (see details) [default: "mean"].}

\item{keep.cut}{A boolean, indicating if the last window must be kept if it is uncompleted [default: TRUE].}
}
\description{
A fast and simple sliding window tool for R.
}
