% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sliding_window_mean}
\alias{sliding_window_mean}
\title{Sliding window mean}
\usage{
sliding_window_mean(x, length, by, keep_cut)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{length}{An integer. Length of a window.}

\item{by}{An integer. Step between each window.}

\item{keep_cut}{A boolean. Indicate if the last uncompleted window should be kept or not.}
}
\description{
Compute mean of each sliding window.
}
